azure-cost-control/requirements.txt
pulumi==3.35.0
pulumi-azure==5.31.0
pulumi-azuread==5.34.0
pulumi-azure-native==1.92.0
